+++ js/main.js
@@ -145,6 +145,8 @@ function saveAll(){
             console.log(errorThrown);
         }
     });
+	
+	calculateStats();
 }
 
 function refreshAll(){
@@ -161,6 +163,7 @@ function refreshAll(){
             $('div.command-btns>button').prop('disabled',false).parent().find('#refreshBtn>span').removeClass('gly-spin');
             formatTooltips();
 			callBackTime();
+			calculateStats();
         }
     });
 }
@@ -181,7 +184,7 @@ function callBackTime(){
 					var dayMon = word.split('/');
 					day = dayMon[0];
 					mon = dayMon[1];
-					hou = (typeof dayMon[2]!== 'undefined' ? dayMon[2] : false);
+					hou = (typeof dayMon[2]!== 'undefined' ? dayMon[2].toUpperCase() : false);
 
 					isDate = ($.isNumeric(day)&&$.isNumeric(mon) ? true : false);
 				}
@@ -212,6 +215,10 @@ function formatTooltips(){
     $('[data-toggle="tooltip"]').tooltip({
         placement : 'top'
     });
+	
+	$('[data-toggle="tooltip_below"]').tooltip({
+		placement : 'right'
+	});
 }
 
 function openCode(){
@@ -322,4 +329,146 @@ function autoRefresh(){
 	btn.attr('data-checked')=='1'
 			? btn.attr('data-checked','0').find('span').removeClass('glyphicon-check').addClass('glyphicon-unchecked')
 			: btn.attr('data-checked','1').find('span').removeClass('glyphicon-unchecked').addClass('glyphicon-check');
+}
+
+function calculateStats(){
+	//tasks-t, tasks-y, tasks-a : tasks #, % over yesterday/average
+	//bd-t, bd-y, bd-a : business dev #, % over yesterday/average
+	//bd>vb>em>cb>lm>rb>ni% today/yesterday/all-time tasks
+	
+	var d = new Date();
+	var tmon = d.getMonth()+1;
+	var tday = d.getDate();
+	
+	ymon = tmon;
+	
+	if(tday-1==0){
+		tmon-1==0 ? ymon=12 : ymon=tmon-1;
+		if(ymon==1||ymon==3||ymon==5||ymon==7||ymon==8||ymon==10||ymon==12) yday=31;
+		if(ymon==4||ymon==6||ymon==9||ymon==11)	yday=30;
+		if(ymon==2) yday=28;
+	} else yday=tday-1;
+	
+	//Tasks	
+	var tasks_t = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]').length;
+	var tasks_y = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]').length;
+	var tasks_a = $('div#rows>div.column>div.row').length;	
+	$('#stat-tasks-t').html(tasks_t);
+	$('#stat-tasks-y').html(tasks_y);
+	$('#stat-tasks-a').html(tasks_a);
+	
+	//Business Development
+	var bd_t = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Business Development"]:not(.btn-default)').length+$('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Not Interested"]:not(.btn-default)').length;
+	var bd_y = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Business Development"]:not(.btn-default)').length+$('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Not Interested"]:not(.btn-default)').length;
+	var bd_c = Math.round((bd_t-bd_y)/bd_y*100);
+	$('#stat-bd-t').html(bd_t+' ('+Math.round((bd_t/tasks_t)*100)+'%)');
+	$('#stat-bd-y').html(bd_y+' ('+Math.round((bd_y/tasks_y)*100)+'%)');
+	$('#stat-bd-c').html((!isNaN(bd_c) ? (bd_t-bd_y) : (bd_t-bd_y)+' ('+bd_c+'%)'));
+	
+	//Graphs
+		//Today
+		var g_t_bd = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Business Development"]:not(.btn-default)').length;
+		var g_t_vb = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Visit Booked"]:not(.btn-default)').length;
+		var g_t_se = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Sent Email"]:not(.btn-default)').length;
+		var g_t_cl = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Call Later"]:not(.btn-default)').length;
+		var g_t_lm = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Left Message"]:not(.btn-default)').length;
+		var g_t_ni = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Not Interested"]:not(.btn-default)').length;
+		var g_t_rb = $('div.row[data-date-month="'+tmon+'"][data-date-day="'+tday+'"]>div.input-group>div.task-btns>button[data-original-title="Reception Block/Couldn\'t Connect"]:not(.btn-default)').length;
+		//As percentages
+		g_t_bd_p = (g_t_bd/tasks_t*100);
+		g_t_vb_p = (g_t_vb/tasks_t*100);
+		g_t_se_p = (g_t_se/tasks_t*100);
+		g_t_cl_p = (g_t_cl/tasks_t*100);
+		g_t_lm_p = (g_t_lm/tasks_t*100);
+		g_t_ni_p = (g_t_ni/tasks_t*100);
+		g_t_rb_p = (g_t_rb/tasks_t*100);
+		g_t_total = g_t_bd_p+g_t_vb_p+g_t_se_p+g_t_cl_p+g_t_lm_p+g_t_ni_p+g_t_rb_p;		
+		
+		//Output
+		$('#stat-g-t>.progress>.bd').css('width',g_t_bd_p+'%');
+		$('#stat-g-t>.progress>.vb').css('width',g_t_vb_p+'%');
+		$('#stat-g-t>.progress>.se').css('width',g_t_se_p+'%');
+		$('#stat-g-t>.progress>.cl').css('width',g_t_cl_p+'%');
+		$('#stat-g-t>.progress>.lm').css('width',g_t_lm_p+'%');
+		$('#stat-g-t>.progress>.ni').css('width',g_t_ni_p+'%');
+		$('#stat-g-t>.progress>.rb').css('width',g_t_rb_p+'%');
+		$('#stat-gl-t-bd').html((Math.round(g_t_bd_p))+'%');
+		$('#stat-gl-t-vb').html((Math.round(g_t_vb_p))+'%');
+		$('#stat-gl-t-se').html((Math.round(g_t_se_p))+'%');
+		$('#stat-gl-t-cl').html((Math.round(g_t_cl_p))+'%');
+		$('#stat-gl-t-lm').html((Math.round(g_t_lm_p))+'%');
+		$('#stat-gl-t-ni').html((Math.round(g_t_ni_p))+'%');
+		$('#stat-gl-t-rb').html((Math.round(g_t_rb_p))+'%');
+		$('#stat-gl-t-in').html(Math.round(100-g_t_total)+'%');
+		
+		//Yesterday
+		var g_y_bd = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Business Development"]:not(.btn-default)').length;
+		var g_y_vb = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Visit Booked"]:not(.btn-default)').length;
+		var g_y_se = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Sent Email"]:not(.btn-default)').length;
+		var g_y_cl = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Call Later"]:not(.btn-default)').length;
+		var g_y_lm = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Left Message"]:not(.btn-default)').length;
+		var g_y_ni = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Not Interested"]:not(.btn-default)').length;
+		var g_y_rb = $('div.row[data-date-month="'+ymon+'"][data-date-day="'+yday+'"]>div.input-group>div.task-btns>button[data-original-title="Reception Block/Couldn\'t Connect"]:not(.btn-default)').length;
+		//As percentages
+		g_y_bd_p =(g_y_bd/tasks_y*100);
+		g_y_vb_p =(g_y_vb/tasks_y*100);
+		g_y_se_p =(g_y_se/tasks_y*100);
+		g_y_cl_p =(g_y_cl/tasks_y*100);
+		g_y_lm_p =(g_y_lm/tasks_y*100);
+		g_y_ni_p =(g_y_ni/tasks_y*100);
+		g_y_rb_p =(g_y_rb/tasks_y*100);
+		g_y_total = g_y_bd_p+g_y_vb_p+g_y_se_p+g_y_cl_p+g_y_lm_p+g_y_ni_p+g_y_rb_p;		
+		
+		//Output
+		$('#stat-g-y>.progress>.bd').css('width',g_y_bd_p+'%');
+		$('#stat-g-y>.progress>.vb').css('width',g_y_vb_p+'%');
+		$('#stat-g-y>.progress>.se').css('width',g_y_se_p+'%');
+		$('#stat-g-y>.progress>.cl').css('width',g_y_cl_p+'%');
+		$('#stat-g-y>.progress>.lm').css('width',g_y_lm_p+'%');
+		$('#stat-g-y>.progress>.ni').css('width',g_y_ni_p+'%');
+		$('#stat-g-y>.progress>.rb').css('width',g_y_rb_p+'%');
+		$('#stat-gl-y-bd').html((Math.round(g_y_bd_p))+'%');
+		$('#stat-gl-y-vb').html((Math.round(g_y_vb_p))+'%');
+		$('#stat-gl-y-se').html((Math.round(g_y_se_p))+'%');
+		$('#stat-gl-y-cl').html((Math.round(g_y_cl_p))+'%');
+		$('#stat-gl-y-lm').html((Math.round(g_y_lm_p))+'%');
+		$('#stat-gl-y-ni').html((Math.round(g_y_ni_p))+'%');
+		$('#stat-gl-y-rb').html((Math.round(g_y_rb_p))+'%');
+		$('#stat-gl-y-in').html(Math.round(100-g_y_total)+'%');
+		
+		//All-Time
+		var g_a_bd = $('div.row>div.input-group>div.task-btns>button[data-original-title="Business Development"]:not(.btn-default)').length;
+		var g_a_vb = $('div.row>div.input-group>div.task-btns>button[data-original-title="Visit Booked"]:not(.btn-default)').length;
+		var g_a_se = $('div.row>div.input-group>div.task-btns>button[data-original-title="Sent Email"]:not(.btn-default)').length;
+		var g_a_cl = $('div.row>div.input-group>div.task-btns>button[data-original-title="Call Later"]:not(.btn-default)').length;
+		var g_a_lm = $('div.row>div.input-group>div.task-btns>button[data-original-title="Left Message"]:not(.btn-default)').length;
+		var g_a_ni = $('div.row>div.input-group>div.task-btns>button[data-original-title="Not Interested"]:not(.btn-default)').length;
+		var g_a_rb = $('div.row>div.input-group>div.task-btns>button[data-original-title="Reception Block/Couldn\'t Connect"]:not(.btn-default)').length;
+		//As percentages
+		g_a_bd_p = (g_a_bd/tasks_a*100);
+		g_a_vb_p = (g_a_vb/tasks_a*100);
+		g_a_se_p = (g_a_se/tasks_a*100);
+		g_a_cl_p = (g_a_cl/tasks_a*100);
+		g_a_lm_p = (g_a_lm/tasks_a*100);
+		g_a_ni_p = (g_a_ni/tasks_a*100);
+		g_a_rb_p = (g_a_rb/tasks_a*100);
+		g_a_total = g_a_bd_p+g_a_vb_p+g_a_se_p+g_a_cl_p+g_a_lm_p+g_a_ni_p+g_a_rb_p;
+		//Output
+		$('#stat-g-a>.progress>.bd').css('width',g_a_bd_p+'%');
+		$('#stat-g-a>.progress>.vb').css('width',g_a_vb_p+'%');
+		$('#stat-g-a>.progress>.se').css('width',g_a_se_p+'%');
+		$('#stat-g-a>.progress>.cl').css('width',g_a_cl_p+'%');
+		$('#stat-g-a>.progress>.lm').css('width',g_a_lm_p+'%');
+		$('#stat-g-a>.progress>.ni').css('width',g_a_ni_p+'%');
+		$('#stat-g-a>.progress>.rb').css('width',g_a_rb_p+'%');
+		$('#stat-gl-a-bd').html((Math.round(g_a_bd_p))+'%');
+		$('#stat-gl-a-vb').html((Math.round(g_a_vb_p))+'%');
+		$('#stat-gl-a-se').html((Math.round(g_a_se_p))+'%');
+		$('#stat-gl-a-cl').html((Math.round(g_a_cl_p))+'%');
+		$('#stat-gl-a-lm').html((Math.round(g_a_lm_p))+'%');
+		$('#stat-gl-a-ni').html((Math.round(g_a_ni_p))+'%');
+		$('#stat-gl-a-rb').html((Math.round(g_a_rb_p))+'%');
+		$('#stat-gl-a-in').html(Math.round(100-g_a_total)+'%');
+		
+		formatTooltips();
 }