+++ sessions/chats/createlog.php
@@ -1,0 +1,10 @@
+<form action="" method="post">
+	<textarea name="content"></textarea>
+	<input type="submit" value="submit" />
+</form>
+
+<?php
+
+if(isset($_POST['content'])){
+	
+}